Power of 10 function is corrected

+++ src/float_spec.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   float_spec.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yorazaye <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: yorazaye <yorazaye@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/02 23:38:09 by yorazaye          #+#    #+#             */
-/*   Updated: 2019/11/02 23:41:22 by yorazaye         ###   ########.fr       */
+/*   Updated: 2019/11/03 14:10:56 by yorazaye         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@ int			f_spec(va_list va)
 {
 	int		r;
 
-	r = ft_putfloat((float)va_arg(va, double), 6);
+	r = ft_putfloat((double)va_arg(va, double), 6);
 	return (r);
 }
