+++ src/ft_putfloat.c
@@ -6,7 +6,7 @@
 /*   By: yorazaye <yorazaye@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/01 17:19:37 by yorazaye          #+#    #+#             */
-/*   Updated: 2019/11/02 23:34:06 by yorazaye         ###   ########.fr       */
+/*   Updated: 2019/11/03 14:09:45 by yorazaye         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 static double	ft_roundup(double n, int afterpoint)
 {
-	double	pwr;
-	double	numbs;
+	long double	pwr;
+	double		numbs;
 
 	pwr = ft_pow(10, afterpoint + 1);
 	numbs = n * pwr - (double)((int)(n * pwr / 10) * 10);
